Refactor AI to split the ai player from other functions.
