+++ src/ai/mod.rs
@@ -1,2 +1,4 @@
 pub mod ai_utils;
 pub mod ai_action;
+pub mod ai_play_card;
+pub mod ai_player;
